+++ CHANGELOG.md
@@ -1,5 +1,26 @@
 # Changelog
 
+## Version 0.0.3
+
+* KONTAKT 5.6.5 ready. Following command, variable added
+
+    - New UI control
+        - ui_xy
+        - $CONTROL_PAR_CURSOR_PICTURE
+        - $CONTROL_PAR_MOUSE_MODE
+        - $CONTROL_PAR_ACTIVE_INDEX
+        - $CONTROL_PAR_MOUSE_BEHAVIOUR_X
+        - $CONTROL_PAR_MOUSE_BEHAVIOUR_Y
+
+	- New UI Commands
+        - New UI commands: set_control_par_arr()
+        - set_control_par_str_arr()
+
+* Improved
+
+    - KSP commands signature behavior  (hover)
+    - Syntax highlighting (KSP commands)
+
 ## Version 0.0.2
 
 * Autocomplete supported